# Conflicts:
#	src/EditorFeatures/Core.Wpf/Microsoft.CodeAnalysis.EditorFeatures.Wpf.csproj
#	src/VisualStudio/CSharp/Impl/Options/AdvancedOptionPageControl.xaml
#	src/VisualStudio/CSharp/Impl/Options/AdvancedOptionPageStrings.cs
#	src/VisualStudio/CSharp/Impl/PackageRegistration.pkgdef
#	src/VisualStudio/CSharp/Impl/UnifiedSettings/csharpSettings.registration.json
#	src/VisualStudio/Core/Def/ServicesVSResources.resx
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.cs.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.de.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.es.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.fr.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.it.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.ja.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.ko.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.pl.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.pt-BR.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.ru.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.tr.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.zh-Hans.xlf
#	src/VisualStudio/Core/Def/xlf/ServicesVSResources.zh-Hant.xlf
#	src/VisualStudio/Core/Test.Next/UnifiedSettings/UnifiedSettingsTests.cs
#	src/VisualStudio/VisualBasic/Impl/Options/AdvancedOptionPageControl.xaml
#	src/VisualStudio/VisualBasic/Impl/Options/AdvancedOptionPageStrings.vb

+++ src/Features/Core/Portable/ReplaceMethodWithProperty/ReplaceMethodWithPropertyCodeRefactoringProvider.cs
@@ -385,7 +385,7 @@ internal sealed class ReplaceMethodWithPropertyCodeRefactoringProvider() : CodeR
             var setMethodDeclaration = await GetMethodDeclarationAsync(setMethod, cancellationToken).ConfigureAwait(false);
 
             var setMethodDocument = updatedSolution.GetDocument(setMethodDeclaration?.SyntaxTree);
-            if (setMethodDocument?.Id == documentId)
+            if (setMethodDocument?.Id == documentId && setMethodDeclaration != null)
             {
                 service.RemoveSetMethod(editor, setMethodDeclaration);
             }
